requirements.txt
Django==3.2.6
django-environ==0.4.5
django-debug-toolbar==3.2.2
psycopg2-binary==2.9.1

djangorestframework==3.12.4
django-rest-swagger==2.2.0
django-allauth==0.45.0
dj-rest-auth==2.1.11

requests==2.26.0

pytest==6.2.4
pytest-django==4.4.0
pytest-cov==2.12.1
pytest-xdist==2.3.0
coverage==5.5

flake8==3.9.2
flake8-coding==1.3.2
flake8-quotes==3.3.0
flake8-isort==4.0.0

